+++ CHANGELOG.md
@@ -7,10 +7,15 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+### Added
+
+* \[Markdown] Support for jira quick refs. Ref with format `[JIRA:ABC-123]` will insert jira issue reference macro if
+  reference is not defined in document explicitly (#20)
+
 ### Fixed
 
 * Don't fail when modifying/deleting attachment in Confluence Server (#18)
-* User reference now supports `-` symbol in the middle of username, e.g. `~user-name` (#21)
+* \[Markdown] User reference now supports `-` symbol in the middle of username, e.g. `~user-name` (#21)
 
 ## 0.4.1 - 2022-03-14